requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10

